+++ modules/oit/include/modules/oit/processors/rasterizer.h
@@ -76,7 +76,7 @@ public:
      * @param imageSize Size in pixels.
      * @param worldMatrixTransform Additional transform to be applied before rendering.
      */
-    virtual void rasterize(const ivec2& imageSize, const mat4& worldMatrixTransform) = 0;
+    virtual void rasterize(const ivec2& imageSize, const dmat4& worldMatrixTransform) = 0;
 
     /**
      * @brief Query whether fragments will be emitted.
