+++ modules/oit/include/modules/oit/processors/volumerasterizer.h
@@ -51,7 +51,7 @@ public:
 
     virtual void initializeResources() override;
 
-    virtual void rasterize(const ivec2& imageSize, const mat4& worldMatrixTransform) override;
+    virtual void rasterize(const ivec2& imageSize, const dmat4& worldMatrixTransform) override;
 
     virtual UseFragmentList usesFragmentLists() const override {
         return VolumeFragmentListRenderer::supportsFragmentLists() ? UseFragmentList::Yes
